fix the following warning, introduced in fix the following warning, introduced in r5487: {{{ inode.c:3224: warning: format "%lld" expects type "long long int", but argument 11 has type "gfarm_int64_t" }}}
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@5488 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ server/gfmd/inode.c
@@ -3227,7 +3227,7 @@ inode_replicated(struct file_replicating *fr,
 			    (long long)inode_get_number(inode),
 			    (long long)fr->igen,
 			    host_name(fr->dst),
-			    inode->i_mode, inode_get_gen(inode));
+			    inode->i_mode, (long long)inode_get_gen(inode));
 		} else if (
 		    src_errcode != GFARM_ERR_NO_ERROR ||
 		    dst_errcode != GFARM_ERR_NO_ERROR)
